backend/fastapi/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
python-dotenv==1.0.1
psycopg2-binary
pytest
pytest-mock

